Avoid multiple defination errors.
